move some function from metric_utils to mmedit/utils
