+++ mmedit/evaluation/metrics/niqe.py
@@ -11,8 +11,8 @@ from scipy.special import gamma
 
 from mmedit.datasets.transforms import MATLABLikeResize
 from mmedit.registry import METRICS
+from mmedit.utils import reorder_image, to_numpy
 from .base_sample_wise_metric import BaseSampleWiseMetric
-from .metrics_utils import reorder_image, to_numpy
 
 
 @METRICS.register_module()
