+++ mmedit/utils/__init__.py
@@ -1,11 +1,11 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .cli import modify_args
+from .img_utils import reorder_image, tensor2img, to_numpy
 from .io_utils import MMGEN_CACHE_DIR, download_from_url
 # TODO replace with engine's API
 from .logger import print_colored_log
 from .sampler import get_sampler
 from .setup_env import register_all_modules
-from .tensor2img import tensor2img
 from .trans_utils import (add_gaussian_noise, adjust_gamma, bbox2mask,
                           brush_stroke_mask, get_irregular_mask, make_coord,
                           random_bbox, random_choose_unknown)
@@ -16,5 +16,6 @@ __all__ = [
     'ForwardInputs', 'SampleList', 'NoiseVar', 'LabelVar', 'MMGEN_CACHE_DIR',
     'download_from_url', 'get_sampler', 'tensor2img', 'random_choose_unknown',
     'add_gaussian_noise', 'adjust_gamma', 'make_coord', 'bbox2mask',
-    'brush_stroke_mask', 'get_irregular_mask', 'random_bbox'
+    'brush_stroke_mask', 'get_irregular_mask', 'random_bbox', 'reorder_image',
+    'to_numpy'
 ]
