+++ tests/test_utils/test_tensor2img.py
@@ -1,8 +1,9 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+import numpy as np
 import pytest
 import torch
 
-from mmedit.utils import tensor2img
+from mmedit.utils import tensor2img, to_numpy
 
 
 def test_tensor2img():
@@ -29,3 +30,9 @@ def test_tensor2img():
 
     with pytest.raises(ValueError):
         tensor2img([torch.randn(2, 3, 4, 4, 4) for _ in range(2)])
+
+
+def test_to_numpy():
+    input = torch.rand(1, 3, 8, 8)
+    output = to_numpy(input)
+    assert isinstance(output, np.ndarray)
